airflow_streaming/requirements.txt
apache-airflow-providers-postgres
requests
pandas
psycopg2-binary